+++ PathFindingTest/HPAStar.h
@@ -9,7 +9,7 @@ private:
 	{
 		Vec2D _position;
 		HPANode** _internalNodes;		//pointers to all nodes within the cluster. All nodes are placed along the edges. cap = 2* _clusterSize
-		float** _internalPathLengths;		//table of path length between each internal node
+		double** _internalPathLengths;		//table of path length between each internal node
 		int _nrOfInternalNodes;
 		int _nodeCap;
 		Cluster()
@@ -48,7 +48,7 @@ private:
 	void findEdges(Vec2D pos, const Vec2D dir, Cluster* cluster1, Cluster* cluster2, Metrics& metrics);		//dir = along the edge from pos. symm(t) = t + {dir.y, dir.x}
 	void setEdgePair(Vec2D pos1, Vec2D pos2, Cluster* cluster1, Cluster* cluster2, Metrics& metrics);
 	void findInternalPaths(Cluster* cluster, Metrics& metrics);
-	float* attachNodeToGraph(HPANode* node, Metrics& metrics);				//Used to fix start and goal nodes to the high level graph
+	double* attachNodeToGraph(HPANode* node, Metrics& metrics);				//Used to fix start and goal nodes to the high level graph
 	void calculateGCost(HPANode* parentNode, HPANode* childNode, int distance);
 	void calculateHCost(HPANode* node);
 public:
